tianchi: add xhdpi elements
tianchi is a hdpi phablet, but its 720x1280 resolution means
that many apps treat it like an xhdpi phone. Include the higher
resolution elements so we don't have blurry icons in those
situations. It adds less than 10mb to the final ROM and we
have plenty of room in /system.

Change-Id: Ib5c746234617150978bb87ffc8ddb29bc17d79db
Signed-off-by: Adam Farden <adam@farden.cz>